+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - rendering D&D 2024 spells for 1 level semi spell classes
 - managing Daggerheart domain cards
+- item bonuses for Daggerheart characters
 
 ## [0.3.10] - 2025-10-22
 ### Added